backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.12.1
CacheControl==0.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
cryptography==46.0.6
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
firebase_admin==7.3.0
google-api-core==2.30.1
google-auth==2.49.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
groq==1.0.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
joblib==1.5.3
msgpack==1.1.2
numpy==2.0.2
opencv-python==4.13.0.92
packaging==26.0
pillow==11.3.0
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
pytesseract==0.3.13
python-dotenv==1.2.1
requests==2.32.5
scikit-learn==1.6.1
scipy==1.13.1
sniffio==1.3.1
sse-starlette==3.3.0
starlette==0.49.3
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
